+++ public/css/settings.css
@@ -48,5 +48,7 @@
 }
 
 .settings.details-inner {
-	max-width: 400px;
+	display: grid;
+	grid-gap: 20px 40px;
+	grid-template-columns: repeat(auto-fill, minmax(400px, 1fr));
 }
