+++ src/routers/index.ts
@@ -2,7 +2,6 @@ import home from "./home";
 import mangaPage from "./manga-page";
 import search from "./search";
 import notFound from "./not-found";
-import backups from "./backups";
 import lists from "./lists";
 import settings from "./settings";
 
@@ -11,7 +10,6 @@ export default {
 	mangaPage,
 	search,
 	notFound,
-	backups,
 	lists,
 	settings
 }
