+++ views/backups.handlebars
@@ -11,5 +11,3 @@
 	</div>
 
 </div>
-
-<script src="/js/backups.js"></script>
