+++ views/settings.handlebars
@@ -24,7 +24,18 @@
 		<h2 class="sub-title">Reader settings</h2>
 		{{> settings-inner }}
 	</section>
+	<section class="section">
+		<h2 class="sub-title">Backups</h2>
+		<div class="backups">
+		{{#each backups}}
+		<div class="backup" data-backup="{{fileName}}">
+			<span>{{label}}</span>
+			<button class="badge-background">Restore</button>
+		</div>
+		{{/each}}
+	</div>
 </div>
 
 <script src="/js/reader-settings.js"></script>
 <script src="/js/settings.js"></script>
+<script src="/js/backups.js"></script>
